removed the proxy for o2 Ireland
Cutomers on MCC 272 MNC 02 should not be using the proxy we have identified 10000
handsets using the proxy and this is down to the Android source code we cannot OTA these
handsets can go directly to the internet with better throughput and less page timeouts
The current proxy is really only used for MMS header enrichment.
I manage the proxy and the traffic is increasing month on month as are customer complaints.
This is why this configuration should be removed.

Test: made changes on rooted handset and verified they did not use the proxy but could browse the internet

Change-Id: I56f54c006cbe81bbf5ccc9a0e0c20885b94f2aba
Signed-off-by: Gareth Evans <garethe@originalsolutions.co.uk>